6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,8 +7,8 @@
 int main(void)
 {
 	char charType;
-	short shortInt;
-	int intType;
+	int shortInt;
+	unsigned int intType;
 	float floatType;
 	double doubleType;
